+++ BBFitnessTrackerAPI/Core/Entities/CalorieData.cs
@@ -15,9 +15,9 @@ namespace Core.Entities
         [ForeignKey(nameof(UserId))]
         public User ? User { get; set; } 
         public DateTime Datum { get; set; }
-        public int Kalorienaufnahme { get; set; }
-        public int Fette { get; set; }
-        public int Proteine { get; set; }
-        public int Kohlenhydrate { get; set; }
+        public double Kalorienaufnahme { get; set; }
+        public double Fette { get; set; }
+        public double Proteine { get; set; }
+        public double Kohlenhydrate { get; set; }
     }
 }
